ACLModuleManager::GetACLModule is not consistent
If it is able to load (i.e. dlopen) the ACL module but fails
to initialize the compiler (i.e. CompilerInit fails), it will
report failure.

However, the next time you call it, it will report success.

This is because is caches the sucessful load, and assumes that
CompilerInit was already called sucessfully.

This change makes sure that everything was successful, not just the module load